Sanitized Inputs
Redid the text input section to allow the user to input words in upper or lower case without us having to check the casing of each character. This also makes it so that when entering options into the menu, you only have to enter the text in uppercase, since all text will be converted to uppercase.

Fixed the HELP message accordingly.